blog/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-classy-tags==3.0.1
django-cookie-law==2.1.1
django-js-asset==2.0.0
django-modeltranslation==0.17.5
mysqlclient==2.1.0
Pillow==9.0.1
python-dotenv==0.19.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
